fix: triton pipeline
